Features complete. Gradle bootRun command line gets stuck at 75%. Works from IntelliJ.
