+++ src/main/java/com/fellowscode/fishersarah/songr/SongrApplication.java
@@ -4,10 +4,13 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
+
 public class SongrApplication {
 
 	public static void main(String[] args) {
+
 		SpringApplication.run(SongrApplication.class, args);
+
 	}
 
 }
