+++ src/main/java/com/fellowscode/fishersarah/songr/models/Album.java
@@ -1,13 +1,26 @@
-package com.fellowscode.fishersarah.songr;
+package com.fellowscode.fishersarah.songr.models;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
 public class Album {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long Id;
+
     public String title;
     public String artist;
     public int songCount;
     public int length; //in seconds
     public String imageUrl;
 
-    Album(String title, String artist, int songCount, int length, String imageUrl) {
+    public Album(){}
+
+    public Album(String title, String artist, int songCount, int length, String imageUrl) {
         this.title = title;
         this.artist = artist;
         this.songCount = songCount;
