+++ src/main/java/com/fellowscode/fishersarah/songr/models/AlbumRepository.java
@@ -1,0 +1,10 @@
+package com.fellowscode.fishersarah.songr.models;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+//Jpa Repository <Type, Identifier type>
+//Album Id was a long so this needs to be a long
+
+public interface AlbumRepository extends JpaRepository<Album, Long> {
+
+}
