+++ src/main/resources/application.properties
@@ -1,1 +1,6 @@
-
+spring.datasource.url=jdbc:postgresql://localhost:5432/albums
+spring.datasource.username=sajafisher
+spring.datasource.password=willcode
+spring.jpa.hibernate.ddl-auto=update
+# ddl-auto=create wipes the data on reload
+# ddl-auto=update only updates, doesn't wipe
